refactor(proxy): extract CheckTarget adapter out of runner.py
Move _entry_to_check_target (267 lines) and the _DANGEROUS_SEGMENTS
constant it uses out of runner.py into a dedicated, dependency-light
dast/proxy/check_target_adapter.py. The adapter is the pure request-shaping
layer between a ProxyEntry and the scanner's CheckTarget — it holds no scan
state and sends no requests, so it does not belong in the runner.

Both symbols are re-imported into runner.py so the scan worker and the
existing test (tests/unit/test_active_checks.py imports _entry_to_check_target
from dast.proxy.runner) keep working unchanged. Behavior-preserving:
code moved verbatim, no logic changes. runner.py drops from 1680 to 1400 lines.
